refactor(station,line): 코드 리팩터링
- dto, entity convert 메소드 생성
- 서비스 비즈니스로직 제거
