+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,6 +1,7 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.common.BaseEntity;
+import nextstep.subway.common.entity.BaseEntity;
+import nextstep.subway.line.dto.LineResponse;
 
 import javax.persistence.*;
 
@@ -19,6 +20,10 @@ public class Line extends BaseEntity {
     public Line(String name, String color) {
         this.name = name;
         this.color = color;
+    }
+
+    public LineResponse toDto() {
+        return LineResponse.of(id, name, color, getCreatedDate(), getModifiedDate());
     }
 
     public void update(Line line) {
