+++ src/main/java/nextstep/subway/station/dto/StationResponse.java
@@ -1,7 +1,5 @@
 package nextstep.subway.station.dto;
 
-import nextstep.subway.station.domain.Station;
-
 import java.time.LocalDateTime;
 
 public class StationResponse {
@@ -20,8 +18,8 @@ public class StationResponse {
         this.modifiedDate = modifiedDate;
     }
 
-    public static StationResponse of(Station station) {
-        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(), station.getModifiedDate());
+    public static StationResponse of(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        return new StationResponse(id, name, createdDate, modifiedDate);
     }
 
     public Long getId() {
